+++ frontend/src/components/map/PoliticalMap.vue
@@ -649,17 +649,6 @@ export default {
         Sort.stringPropAlphabetically('shortName')
       );
 
-      // We remove the heirs from the list if timeline is deactivated,
-      // as they are not displayed there.
-      if (this.timelineActive == false) {
-        this.availableRulers.filter(p => {
-          const role = p.role
-          if (role) {
-            return role.name !== "heir"
-          } else return true
-        })
-      }
-
       /**
        * If a ruler is selected but not in the timeline anymore:
        */